(BSR)[API] refactor: move url-generation functions to utils.urls
