+++ api/src/pcapi/utils/urls.py
@@ -2,6 +2,11 @@ from urllib.parse import urlencode
 
 from pcapi import settings
 from pcapi.core.bookings.models import Booking
+from pcapi.core.educational.models import CollectiveOffer
+from pcapi.core.educational.models import CollectiveOfferTemplate
+from pcapi.core.offerers import models as offerers_models
+from pcapi.core.offers.models import Offer
+from pcapi.utils.human_ids import humanize
 
 
 def generate_firebase_dynamic_link(path: str, params: dict | None) -> str:
@@ -15,3 +20,27 @@ def generate_firebase_dynamic_link(path: str, params: dict | None) -> str:
 
 def booking_app_link(booking: Booking) -> str:
     return f"{settings.WEBAPP_V2_URL}/reservation/{booking.id}/details"
+
+
+def build_pc_pro_offer_link(offer: CollectiveOffer | CollectiveOfferTemplate | Offer) -> str:
+    if isinstance(offer, CollectiveOffer):
+        return f"{settings.PRO_URL}/offre/{humanize(offer.id)}/collectif/edition"
+
+    if isinstance(offer, CollectiveOfferTemplate):
+        return f"{settings.PRO_URL}/offre/T-{humanize(offer.id)}/collectif/edition"
+
+    return f"{settings.PRO_URL}/offre/{humanize(offer.id)}/individuel/edition"
+
+
+def build_pc_pro_offerer_link(offerer: offerers_models.Offerer) -> str:
+    return f"{settings.PRO_URL}/accueil?structure={humanize(offerer.id)}"
+
+
+def build_pc_pro_venue_link(venue: offerers_models.Venue) -> str:
+    if venue.isVirtual:
+        return build_pc_pro_offerer_link(venue.managingOfferer)
+    return f"{settings.PRO_URL}/structures/{humanize(venue.managingOffererId)}/lieux/{humanize(venue.id)}"
+
+
+def build_pc_pro_venue_bookings_link(venue: offerers_models.Venue) -> str:
+    return f"{settings.PRO_URL}/reservations?offerVenueId={humanize(venue.id)}"
